+++ e2e/internal/e2e/imagebuild.go
@@ -1,4 +1,6 @@
-// Copyright (c) 2019-2022, Sylabs Inc. All rights reserved.
+// Copyright (c) 2019-2023, Sylabs Inc. All rights reserved.
+// Copyright (c) Contributors to the Apptainer project, established as
+//   Apptainer a Series of LF Projects LLC.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -54,6 +56,7 @@ type AppDetail struct {
 	Files   []FilePair
 	Install []string
 	Run     []string
+	Start   []string
 	Test    []string
 }